feat: Refactor configuration management by removing obsolete config downloader and integrating it into the worker module
