+++ src/qdash/workflow/entrypoints/handler.py
@@ -15,11 +15,11 @@ from qdash.dbmodel.execution_lock import ExecutionLockDocument
 from qdash.dbmodel.initialize import initialize
 from qdash.workflow.core.calibration.execution_manager import ExecutionManager
 from qdash.workflow.core.calibration.flow import qubex_one_qubit_cal_flow
-from qdash.workflow.integration.config_downloader import update_config
 from qdash.workflow.utils.slack import SlackContents, Status
 from qdash.workflow.utiltask.create_directory import (
     create_directory_task,
 )
+from qdash.workflow.worker.config_downloader.flow import update_config
 from qubex.version import get_package_version
