+++ src/qdash/workflow/utils/task_converter.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict
+from typing import Any
 
 import numpy as np
